Added some function for line intersection in ro-teta coordinates. Also noticed that edge linking through histeresis only helps in some cases.

+++ jni/heap.h
@@ -6,9 +6,9 @@
 #define RIGHT(i) 2*i+1
 
 struct houghLM {
-	int lm; //localmaxim
-	int ro;
-	int teta;
+	float lm; //localmaxim
+	float ro;
+	float teta;
 };
 
 struct Heap {
